Implement QSML for modular loading.
QuickStart Module Loader, or QSML, is the module loader that is used in Nucleus. 0.5 is used as a starting point for Phonon, which will allow us to have disableable modules - though this is down the road and will require bug fixes.

So far, QSML will:

* Scan for and enable modules
* Setup the config file and add defaults
* Provide injection support for modules and classes instantiated by modules

Reload support has also been added.